Something funny.

+++ core/src/main/java/kz/chesschicken/cherrydrupe/jcheck/JavaCheck.java
@@ -30,13 +30,13 @@ import org.jetbrains.annotations.NotNull;
  */
 public class JavaCheck {
 
-    public static void assertJava(@NotNull EnumJavaVersion gotten, byte @NotNull [] okay) {
+    public static void assertJava(@NotNull EnumJavaVersion gotten, int @NotNull [] okay) {
         assertJava(EnumJavaVersion.CURRENT_JAVA_VERSION, gotten, okay);
     }
 
-    public static void assertJava(@NotNull EnumJavaVersion ref, @NotNull EnumJavaVersion gotten, byte @NotNull [] okay) {
-        byte f = EnumJavaVersion.compareJavaVersions(ref, gotten);
-        for(byte a : okay) {
+    public static void assertJava(@NotNull EnumJavaVersion ref, @NotNull EnumJavaVersion gotten, int @NotNull [] okay) {
+        int f = EnumJavaVersion.compareJavaVersions(ref, gotten);
+        for(int a : okay) {
             if(a == f) continue;
             throw new UnsupportedJavaVersionException(ref, gotten);
         }
